+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.6]
+
+-`<khao-image>` can now handle svg images correctly
+
+## [1.6.5]
+
+- added `--khao-image-border-radius` to `<khao-image>`
+
 ## [1.6.4]
 
 - added `customMinWidth` to `<khao-button>`
@@ -314,7 +322,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - added icons to storybook
 - unified handling of icons
-- updated khao-malet to v0.0.9
+- updated khao-malet to v0.0.10
 
 ## [0.0.10]